1st Upload
This is the 1st version of the app.

Use the select region function to ensure that your current conditions are suitable for it to analyse the image. The threshold value is currently the automatic one for the function that it uses, however this can be changed to be flexible in future releases if needed.

The computer vision toolbox is recommended to be downloaded along with some webcam related toolboxes. Download as when the Matlab errors inform you to do so.